requirements.txt
# Core deep learning frameworks
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0

# Data handling and processing
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<3.0.0
Pillow>=8.0.0,<11.0.0

# Visualization and plotting
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.0,<1.0.0
plotly>=5.0.0,<6.0.0

# Machine learning utilities
scikit-learn>=1.0.0,<2.0.0

# Web application (Streamlit)
streamlit>=1.25.0,<2.0.0

# Image processing
opencv-python>=4.5.0,<5.0.0

# Memory optimization
psutil>=5.8.0,<6.0.0